Move get_followers_for_user
